Add yes or no prompts for deleteion of VMs

+++ delete_linked_clones_vbox.sh
@@ -9,11 +9,28 @@
 # for i in {01..04}; do VBoxManage unregistervm ubuntu1604-vm$i --delete; done
 ######################## Junk
 
-# Power off, Delete em'
-for i in {01..04}; do
-  VBoxManage controlvm ubuntu1604-vm$i poweroff;
-  VBoxManage unregistervm ubuntu1604-vm$i --delete;
-done
+echo -n "Delete VMs created from this script? [y|n]: "
+read yno
+case $yno in
+  [yY] | [yY][Ee][Ss] )
+    echo "Here we go!"
 
+    # Power off, Delete em'
+    for i in {01..04}; do
+      VBoxManage controlvm ubuntu1604-vm$i poweroff;
+      VBoxManage unregistervm ubuntu1604-vm$i --delete;
+    done
 
-mv hosts hosts.deleted.$(date +%F.%T)
+    # Save a copy of the hosts file
+    mv hosts hosts.deleted.$(date +%F.%T)
+  ;;
+
+  [nN] | [n|N][O|o] )
+    echo "No action taken.";
+    exit 1
+    ;;
+
+  *) echo "Invalid input"
+    ;;
+
+esac
